your/demo/helloworld/requirements.txt
###############################################################################
# pip dependent packages
###############################################################################

###############################################################################
# default ARGOS-LABS SDK module
###############################################################################
alabs.common

###############################################################################
# For your dependent modules
###############################################################################
# ...
